fix: 로직 분리
