+++ src/components/SpellCheck/index.tsx
@@ -1,7 +1,8 @@
-import React, { useState, Fragment } from 'react';
+import React, { useState, Fragment, useEffect } from 'react';
 import * as _ from './style';
 import { ErrorInfo } from '@/components/Input';
 import { Button } from '@/components/Button';
+import { GetspellCheck } from '@/api/spells';
 
 interface ErrorSegmentProps {
   text: string;
@@ -10,33 +11,93 @@ interface ErrorSegmentProps {
 }
 
 export const ErrorDisplay = ({
+  value,
+  setValue,
   isClick,
-  segments,
-  handleSuggestionClick,
 }: {
+  value: string;
+  setValue: (e: string) => void;
   isClick: boolean;
-  segments: any[];
-  handleSuggestionClick: (index: number) => void;
-}) => (
-  <_.ErrorWrapper style={{ display: isClick ? 'flex' : 'none' }}>
-    {segments.map((segment, index) => (
-      <div key={index}>
-        {segment.error ? (
-          <Fragment>
-            <ErrorSegment
-              text={segment.text}
-              error={segment.error}
-              onClick={() => handleSuggestionClick(index)}
-            />
-          </Fragment>
-        ) : (
-          <div>{segment.text}</div>
-        )}
-      </div>
-    ))}
-  </_.ErrorWrapper>
-);
+  setIsClick: React.Dispatch<React.SetStateAction<boolean>>;
+}) => {
+  const { mutate: GetSpellMutataion, data: rightSpell } = GetspellCheck();
+  const [segments, setSegments] = useState<
+    { text: string; error?: ErrorInfo }[]
+  >([]);
 
+  useEffect(() => {
+    if (rightSpell && rightSpell.errorInfo) {
+      const errorInfo = rightSpell.errorInfo;
+      let lastIndex = 0;
+      const newSegments: { text: string; error?: ErrorInfo }[] = [];
+
+      errorInfo.forEach((error) => {
+        const { start, end, orgStr, candWord } = error;
+        if (start > lastIndex) {
+          newSegments.push({ text: value.slice(lastIndex, start) });
+        }
+        newSegments.push({
+          text: value.slice(start, end),
+          error: { help: error.help, orgStr, candWord },
+        });
+        lastIndex = end;
+      });
+
+      if (lastIndex < value.length) {
+        newSegments.push({ text: value.slice(lastIndex) });
+      }
+
+      setSegments(newSegments);
+    }
+  }, [value, rightSpell]);
+
+  useEffect(() => {
+    if (!isClick) {
+      const updatedSegments = segments.map((segment) =>
+        segment.error ? { text: segment.text } : { text: segment.text }
+      );
+
+      const updatedValue = updatedSegments
+        .map((segment) => segment.text)
+        .join('');
+
+      setValue(updatedValue);
+    } else {
+      GetSpellMutataion(value);
+    }
+  }, [isClick]);
+
+  const handleSuggestionClick = (index: number) => {
+    const updatedSegments = segments.map((segment, i) => {
+      if (i === index && segment.error) {
+        return { text: segment.error.candWord };
+      } else {
+        return { text: segment.text, error: segment.error };
+      }
+    });
+    setSegments(updatedSegments);
+  };
+
+  return (
+    <_.ErrorWrapper style={{ display: isClick ? 'flex' : 'none' }}>
+      {segments.map((segment, index) => (
+        <div key={index}>
+          {segment.error ? (
+            <Fragment>
+              <ErrorSegment
+                text={segment.text}
+                error={segment.error}
+                onClick={() => handleSuggestionClick(index)}
+              />
+            </Fragment>
+          ) : (
+            <div>{segment.text}</div>
+          )}
+        </div>
+      ))}
+    </_.ErrorWrapper>
+  );
+};
 
 export const ErrorSegment = ({ text, error, onClick }: ErrorSegmentProps) => {
   const [hover, setHover] = useState(false);
